科目完工

+++ src/main/java/com/lm/busi/service/SubjectService.java
@@ -8,7 +8,7 @@ import com.lm.busi.model.Subject;
 public interface SubjectService {
     int insertSelective(Subject record);
     
-    void deleteByPrimaryKeys(Integer... ids);
+    void deleteByPrimaryKeys(Short... ids);
     
     int updateByPrimaryKeySelective(Subject record);
     Subject selectByPrimaryKey(Short id);
